On the user page took the table out of the updatepanel, so edits cause a postback, as they should.
Changed the ridedetails page so the table wasn't pulling data directly from an object, instead it now pulls it's object from an objectdatasource. Added the functionality to edit/delete rides, but methods need to be made to sort out how this will be done. For the delete method you will probably just want to disable the available property and keep them in the database.